Apply site review fixes: hero CLS, editor dev crash, URL sync cost, back-nav, a11y
- server-render the hero section and lazy-mount only the css-doodle
  backdrop, removing the landing page's largest layout shift
- ColorPicker creates a fresh Pickr target per mount, fixing the
  StrictMode crash that blanked every editor page in next dev
- editor URL sync uses native history.replaceState instead of
  router.replace (no more RSC refetch per slider step); query params are
  clamped/validated against option bounds
- "Back to gallery" only walks history when the editor was opened from
  the gallery (sessionStorage marker), otherwise navigates normally
- IntersectionObserver callbacks read the latest batched entry
- focus-visible outlines restored (slider thumb, editor header buttons,
  drawer links); contrast fixes for option labels and editor icon
  buttons; sliders and color swatches get distinct accessible names;
  100vw typo -> 100dvh
- footer Product Hunt badge is a lazy img instead of an eager CSS
  background; optimize-images no longer double-encodes; the Vercel
  ignore-build fallback errs toward building on multi-commit pushes;
  e2e updates for the new nav label + a null-href guard
- agent-outputs/codebase-review-2026-07.md: full review report (fixed
  and remaining findings)
