+++ components/Footer/index.tsx
@@ -13,14 +13,25 @@ export default function Footer() {
             <p>Copyright {new Date().getFullYear()}.</p>
 
             {/* The Product Hunt badge used to live in the homepage header; it
-                now sits in the footer. */}
+                now sits in the footer. A lazy <img> (not a CSS background) so
+                the third-party request neither starts before the footer nears
+                the viewport nor delays the page's load event. */}
             <a
               href="https://www.producthunt.com/posts/tabbied?utm_source=badge-top-post-badge&utm_medium=badge&utm_souce=badge-tabbied"
               className={styles.productHuntBadge}
               target="_blank"
               rel="noreferrer"
-              aria-label="Tabbied on Product Hunt"
-            />
+            >
+              {/* eslint-disable-next-line @next/next/no-img-element */}
+              <img
+                src="https://api.producthunt.com/widgets/embed-image/v1/top-post-badge.svg?post_id=279660&theme=light&period=daily"
+                alt="Tabbied on Product Hunt"
+                width={238}
+                height={49}
+                loading="lazy"
+                decoding="async"
+              />
+            </a>
           </Col>
 
           <Col lg={{ span: 4 }} md={{ span: 5 }}>
