+++ components/ValueSlider/index.tsx
@@ -9,6 +9,8 @@ type ValueSliderProps = {
   step: number;
   value: number;
   onChange: (value: number) => void;
+  /** Accessible name for the thumb (e.g. the option's display name). */
+  label?: string;
 };
 
 export default function ValueSlider({
@@ -17,6 +19,7 @@ export default function ValueSlider({
   step,
   value,
   onChange,
+  label,
 }: ValueSliderProps) {
   return (
     <Slider.Root
@@ -31,7 +34,10 @@ export default function ValueSlider({
         <Slider.Control className={styles.control}>
           <Slider.Track className={styles.track}>
             <Slider.Indicator className={styles.indicator} />
-            <Slider.Thumb className={styles.thumb} aria-label="value slider" />
+            <Slider.Thumb
+              className={styles.thumb}
+              aria-label={label ?? 'value slider'}
+            />
           </Slider.Track>
         </Slider.Control>
         <Slider.Value className={styles.value}>
