+++ components/edit-artwork-page/EditArtwork.module.css
@@ -1,8 +1,8 @@
 .pageWrapper {
   display: flex;
   flex-direction: column;
-  height: 100vw;
   height: 100vh;
+  height: 100dvh;
   overflow: hidden;
 }
 
@@ -78,7 +78,7 @@
   align-items: center;
   justify-content: center;
   padding: 0.35rem;
-  color: var(--gray);
+  color: var(--gray-medium);
   background-color: transparent;
   border: none;
   border-radius: 0.3rem;
@@ -105,7 +105,7 @@
   align-items: center;
   justify-content: center;
   padding: 0.3rem 0.55rem;
-  color: var(--gray);
+  color: var(--gray-medium);
   background-color: transparent;
   border: none;
   transition: color 120ms ease, background-color 120ms ease;
