+++ components/edit-artwork-page/EditArtwork.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useEffect, useRef, useState, type ReactNode } from 'react';
-import { useRouter, usePathname, useSearchParams } from 'next/navigation';
+import { usePathname, useSearchParams } from 'next/navigation';
 import dynamic from 'next/dynamic';
 import { Dialog } from '@base-ui-components/react/dialog';
 import { Shuffle, Expand, X, Minus, Plus } from 'lucide-react';
@@ -145,13 +145,13 @@ export default function EditArtwork({ artwork }: { artwork: Artwork }) {
   const maxColors = artwork.colors?.max ?? paletteDefaults.length;
   const defaultColors = artwork.colors?.default ?? paletteDefaults.length;
 
-  const router = useRouter();
   const pathname = usePathname();
   const searchParams = useSearchParams();
 
   // The value an option takes according to the URL, falling back to the
-  // authored default when the param is absent or malformed (a hand-edited URL
-  // must not feed NaN into a slider, which would blank the control).
+  // authored default when the param is absent or malformed, and clamped /
+  // validated against the option's own bounds (a hand-edited URL must not
+  // feed NaN or an out-of-range value into a control, which would blank it).
   const optionFromQuery = (option: ArtworkOption): OptionValue => {
     const queryVal = searchParams.get(option.id);
 
@@ -160,13 +160,31 @@ export default function EditArtwork({ artwork }: { artwork: Artwork }) {
     }
 
     if (typeof option.default === 'number') {
-      const numericVal = Number(queryVal);
+      // Number('') is 0, which would silently pass the NaN guard.
+      const numericVal = queryVal.trim() === '' ? NaN : Number(queryVal);
 
-      return Number.isNaN(numericVal) ? option.default : numericVal;
+      if (Number.isNaN(numericVal)) {
+        return option.default;
+      }
+
+      return Math.min(
+        Math.max(numericVal, option.min ?? -Infinity),
+        option.max ?? Infinity
+      );
     }
 
     if (typeof option.default === 'boolean') {
       return queryVal === 'true';
+    }
+
+    if (option.type === 'ButtonSelectGroup') {
+      // Grid values follow the aspect ratio (their list is dynamic), so they
+      // validate by shape; other groups must match an authored choice.
+      if (option.id === GRID_OPTION_ID) {
+        return /^\d+x\d+$/.test(queryVal) ? queryVal : option.default;
+      }
+
+      return option.options?.includes(queryVal) ? queryVal : option.default;
     }
 
     return queryVal;
@@ -245,7 +263,7 @@ export default function EditArtwork({ artwork }: { artwork: Artwork }) {
 
   // Sync component state FROM the URL search params after mount (back /
   // forward navigation re-renders with new params without remounting). Each
-  // setter is guarded so an echo of our own router.replace is a no-op.
+  // setter is guarded so an echo of our own URL replace is a no-op.
   useEffect(() => {
     const queryPaletteState = paletteStateFromQuery();
 
@@ -340,7 +358,11 @@ export default function EditArtwork({ artwork }: { artwork: Artwork }) {
     });
 
     if (newParams.toString() !== searchParams.toString()) {
-      router.replace(`${pathname}?${newParams.toString()}`, { scroll: false });
+      // Native replaceState (which the App Router keeps in sync with
+      // useSearchParams) instead of router.replace: the page is fully static,
+      // so a router navigation would refetch an identical RSC payload from
+      // the CDN on every knob tweak — a network request per slider step.
+      window.history.replaceState(null, '', `${pathname}?${newParams.toString()}`);
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [seed, palette, colorCount, optionValues, aspectRatio]);
@@ -449,6 +471,7 @@ export default function EditArtwork({ artwork }: { artwork: Artwork }) {
               step={option.step!}
               value={controlValue as number}
               onChange={onChange}
+              label={option.displayName}
             />
           </div>
         );
