+++ components/edit-artwork-page/EditArtworkHeader.module.css
@@ -30,7 +30,11 @@
   height: 2.75rem;
   border-radius: 0.15rem;
   cursor: pointer;
-  outline: none;
+}
+
+.btn:focus-visible {
+  outline: 2px solid var(--blue-medium);
+  outline-offset: 2px;
 }
 
 .btn .reactIcon {
