+++ lib/galleryScroll.ts
@@ -8,6 +8,8 @@
 export const GALLERY_SCROLL_Y = 'tabbied:gallery-scroll-y';
 export const GALLERY_SCROLL_RESTORE = 'tabbied:gallery-scroll-restore';
 
+export const GALLERY_NAVIGATION = 'tabbied:navigated-from-gallery';
+
 // Called from the editor's back control just before navigating, so the gallery
 // knows to restore (rather than reset) its scroll on the next mount.
 export function armGalleryScrollRestore(): void {
@@ -16,5 +18,30 @@ export function armGalleryScrollRestore(): void {
   } catch {
     // sessionStorage can throw in private mode / sandboxed frames; restoring
     // the scroll position is best-effort, so a failure here is non-fatal.
+  }
+}
+
+// Called when a gallery card is clicked, so the editor it opens knows the
+// previous history entry really is the gallery.
+export function markGalleryNavigation(): void {
+  try {
+    sessionStorage.setItem(GALLERY_NAVIGATION, '1');
+  } catch {
+    // Best-effort, same as above.
+  }
+}
+
+// One-shot read of the marker (consumed on the editor's mount). Without it,
+// "Back to gallery" must not use history.back() — the previous entry could be
+// anything (a search result, another site) when the editor was deep-linked.
+export function consumeGalleryNavigation(): boolean {
+  try {
+    const marked = sessionStorage.getItem(GALLERY_NAVIGATION) === '1';
+
+    sessionStorage.removeItem(GALLERY_NAVIGATION);
+
+    return marked;
+  } catch {
+    return false;
   }
 }
